% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binscatter.R
\name{binscatter}
\alias{binscatter}
\title{Bin-scatter function}
\usage{
binscatter(data, y, x, bins = 20, discrete = FALSE, scatter = FALSE,
  theme = theme_binscatter, fitline = TRUE, controls = c(),
  absorb = c("0"), clustervars = c("0"), pos = "bottom right")
}
\arguments{
\item{data}{Dataframe containing the x and y variables}

\item{y}{The independent variable y}

\item{x}{The dependent variable x}

\item{bins}{The number of bins}

\item{discrete}{Discrete?}

\item{scatter}{Not sure what this is yet...}

\item{theme}{Optional ggplot theme}

\item{fitline}{Option to add a linear fit through the bin scatter}

\item{controls}{Optional control variables to the binscatter}

\item{clustervars}{Optional cluster variables}
}
\description{
This function does a binscatter for two-dimensional data
}
\keyword{binscatter}
