% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_binscatter.R
\name{theme_binscatter}
\alias{theme_binscatter}
\title{Binscatter Theme}
\usage{
theme_binscatter()
}
\description{
A nice ggplot theme for the bin scatter plots
}
\examples{

}
\keyword{binscatter}
\keyword{theme}
